chore: v2026.1.1 commited

+++ src/main/kotlin/com/kdocer/action/KDocerRemoveAction.kt
@@ -1,21 +1,20 @@
 package com.kdocer.action
 
-import com.intellij.codeInsight.CodeInsightSettings
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.command.WriteCommandAction
-import com.intellij.psi.*
+import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.util.PsiTreeUtil
 import org.jetbrains.kotlin.kdoc.psi.api.KDoc
-import org.jetbrains.kotlin.psi.KtClass
+import org.jetbrains.kotlin.psi.KtDeclaration
 import org.jetbrains.kotlin.psi.KtFile
-import org.jetbrains.kotlin.psi.KtNamedFunction
-import org.jetbrains.kotlin.psi.KtProperty
-
 
 /**
+ * Removes the KDoc of the declaration at the caret. Works whether the caret sits inside
+ * the KDoc itself or anywhere on the declaration it documents.
+ *
  * Created by Godwin on 7/23/2020 9:16 PM.
  *
  * @author : Godwin Joseph Kurinjikattu
@@ -23,82 +22,29 @@ import org.jetbrains.kotlin.psi.KtProperty
  */
 class KDocerRemoveAction : AnAction() {
 
+    override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
+
     override fun update(action: AnActionEvent) {
         super.update(action)
-        val presentation = action.presentation
-
-        val psiFile = action.getData(PlatformDataKeys.PSI_FILE) ?: return
-        if (psiFile !is KtFile || !CodeInsightSettings.getInstance().SMART_INDENT_ON_ENTER) {
-            println("This is not Kotlin file. ")
-
-            presentation.isVisible = false
-            presentation.isEnabled = false
-            return
-        }
-        presentation.isVisible = true
-        presentation.isEnabled = true
+        val enabled = action.getData(PlatformDataKeys.PSI_FILE) is KtFile
+        action.presentation.isVisible = enabled
+        action.presentation.isEnabled = enabled
     }
 
     override fun actionPerformed(action: AnActionEvent) {
-        val psiFile = action.getData(PlatformDataKeys.PSI_FILE) ?: return
-        if (psiFile !is KtFile || !CodeInsightSettings.getInstance().SMART_INDENT_ON_ENTER) {
-            println("This is not Kotlin file. ")
-            return
-        }
-        if (!isAllowedElementType(psiFile)) return
-
-        val project = psiFile.project
-        val documentManager = PsiDocumentManager.getInstance(project)
-        documentManager.commitAllDocuments()
-
+        val psiFile = action.getData(PlatformDataKeys.PSI_FILE) as? KtFile ?: return
         val editor = action.getData(PlatformDataKeys.EDITOR) ?: return
-        val caretModel = editor.caretModel
-        val elementAtCaret = psiFile.findElementAt(caretModel.offset) ?: return
+        val project = psiFile.project
+        PsiDocumentManager.getInstance(project).commitAllDocuments()
 
-        preProcessElement(elementAtCaret)
-    }
+        val elementAtCaret = psiFile.findElementAt(editor.caretModel.offset) ?: return
+        // Caret inside the KDoc -> that KDoc; otherwise the enclosing declaration's KDoc.
+        val kdoc = PsiTreeUtil.getParentOfType(elementAtCaret, KDoc::class.java)
+            ?: PsiTreeUtil.getParentOfType(elementAtCaret, KtDeclaration::class.java)?.docComment
+            ?: return
 
-    private fun preProcessElement(element: PsiElement): Boolean {
-        return when (element) {
-            is KDoc -> {
-                processElement(element)
-                true
-            }
-            is PsiWhiteSpace -> {
-                val kDoc = PsiTreeUtil.getParentOfType(element, KDoc::class.java)
-                return if (kDoc != null) {
-                    processElement(kDoc)
-                    true
-                } else {
-                    var status = preProcessElement(element.nextSibling)
-                    if (!status) status = preProcessElement(element.prevSibling)
-                    status
-                }
-            }
-            else -> false
+        WriteCommandAction.runWriteCommandAction(project) {
+            if (kdoc.isValid) kdoc.delete()
         }
-    }
-
-    private fun processElement(kdoc: KDoc) {
-        ApplicationManager.getApplication().invokeLater {
-            WriteCommandAction.runWriteCommandAction(kdoc.project) {
-                kdoc.delete()
-            }
-        }
-    }
-
-    private fun isAllowedElementType(element: PsiElement): Boolean {
-        var result = false
-        if (element is PsiClass ||
-            element is PsiField ||
-            element is PsiMethod ||
-            element is KtFile ||
-            element is KtNamedFunction ||
-            element is KtClass ||
-            element is KtProperty
-        ) {
-            result = true
-        }
-        return result
     }
 }
